requirements.txt
numpy==1.18.5
cython==0.29.23
opencv-python==4.3.0.36
typeguard==2.9.1
pyyaml==5.3.1
plyfile==0.7.2
imgaug==0.4.0
tqdm==4.48.0
progressbar2==3.53.1
h5py==2.10.0
tensorflow-estimator==1.15.1
tensorboard==1.15.0
future==0.18.2
matplotlib==3.4.3
